+++ build-knet.sh
@@ -1,5 +1,5 @@
 
-echo "Building knet"
+echo -e "${COLOR_GREEN}Building knet${COLOR_END}"
 
 KNET_SOURCE_DIR=knet
 if [ ! -d ${KNET_SOURCE_DIR} ] ; then
@@ -9,23 +9,29 @@ else
     echo "-- Repository already cloned, skipping."
 fi
 
-pushd "${KNET_SOURCE_DIR}"
+DETECTION_LIB=${PREFIX}/lib/libkNet.a
+if [ ! -f ${DETECTION_LIB} ] ; then
+    pushd "${KNET_SOURCE_DIR}"
+    
+    export BOOST_ROOT=${PREFIX}
 
-export BOOST_ROOT=${PREFIX}
+    mkdir -p build-android
+    cd build-android
 
-mkdir -p build-android
-cd build-android
+    ${CMAKE_ANDROID} ..
 
-${CMAKE_ANDROID} ..
+    make -j4
 
-make -j4
+    # Manual install, no make install for knet
+    echo "-- Installing libs and headers to ${PREFIX}"
+    cd ..
+    cp -u lib/libkNet.a ${PREFIX}/lib/
+    cp -u -r include/* ${PREFIX}/include/
+    
+    popd
+else
+    echo "-- Already built, remove ${DETECTION_LIB} to trigger a rebuild."
+fi
 
-# Manual install, no make install for knet
-echo "-- Installing libs and headers to ${PREFIX}"
-cd ..
-cp -u lib/*.a ${PREFIX}/lib/
-cp -u -r include/* ${PREFIX}/include/
-
-popd
 echo
